Extract by per-project
This is a more meaningful boundary than the workspace object. AFAICT the per-project bit shouldn't be a slice as per Redux's design.
